+++ packages/gamut-illustrations/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-illustrations@0.10.10...@codecademy/gamut-illustrations@0.11.0) (2021-04-30)
+
+
+### Features
+
+* add new BinaryBlocks illustration ([6ef07e4](https://github.com/Codecademy/client-modules/commit/6ef07e4e6dcffc828bfcff09940bd89d06b424ba))
+
+
+
 ### [0.10.10](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-illustrations@0.10.9...@codecademy/gamut-illustrations@0.10.10) (2021-04-29)
 
 **Note:** Version bump only for package @codecademy/gamut-illustrations